requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.1
annotated-types==0.7.0
antlr4-python3-runtime==4.8
arrow==1.3.0
asttokens==2.4.0
async-generator==1.10
async-timeout==4.0.3
attrs==23.1.0
av==9.2.0
backcall==0.2.0
bddl==1.0.1
beautifulsoup4==4.12.3
black==22.3.0
cachetools==5.3.2
carla==0.9.13
certipy==0.2.1
charset-normalizer==3.3.1
click==8.1.7
colorlog==6.7.0
configparser==7.1.0
contourpy==1.1.1
cycler==0.12.1
datasets==2.14.6
datashape==0.5.4
dill==0.3.7
dm-control==0.0.403778684
dm-env==1.6
dm-tree==0.1.8
docker-pycreds==0.4.0
docopt==0.6.2
easydict==1.9
egl_probe==1.0.2
einops==0.3.2
etils==1.3.0
executing==2.0.0
Farama-Notifications==0.0.4
fastjsonschema==2.20.0
filelock==3.12.4
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.2
gitdb==4.0.11
GitPython==3.1.40
glfw==2.7.0
google==3.0.0
google-auth==2.34.0
google-auth-oauthlib==1.0.0
greenlet==3.0.1
grpcio==1.59.2
gym==0.21.0
gymnasium==0.29.1
h5py==3.11.0
hpack==4.0.0
huggingface-hub==0.18.0
huggingface-sb3==3.0
hydra-core==1.1.1
imageio==2.13.3
imageio-ffmpeg==0.5.1
imitation==1.0.0
importlib-resources==6.1.0
importlib_metadata==8.4.0
iniconfig==2.1.0
ipdb==0.13.9
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonpickle==3.0.2
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterhub==5.2.1
jupyterlab_pygments==0.3.0
jupytext==1.16.7
keyboard==0.13.5
kiwisolver==1.4.5
labmaze==1.0.6
llvmlite==0.41.1
lxml==5.3.0
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
mujoco==3.2.3
multidict==6.0.4
multiprocess==0.70.15
munch==2.5.0
munkres==1.1.4
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.1
nose==1.3.7
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.1.2
opencv-python==4.11.0.86
optuna==3.4.0
packaging==23.2
pamela==1.2.0
pandas==2.0.3
parso==0.8.3
pathspec==0.12.1
pathtools==0.1.2
pickleshare==0.7.5
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
pluggy==1.5.0
polars==1.8.2
promise==2.3
prompt-toolkit==3.0.39
protobuf==3.20.1
psutil==6.0.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybasicbayes==0.2.2
pybullet==3.2.1
pyct==0.4.6
pydantic==2.9.2
pydantic_core==2.23.4
pygame==2.5.2
Pygments==2.16.1
PyOpenGL==3.1.7
pyparsing==3.1.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2023.3.post1
pyzmq==26.2.0
referencing==0.35.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
robomimic==0.3.0
robosuite==1.4.0
rpds-py==0.20.1
rsa==4.9
sacred==0.8.4
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
seals==0.2.1
sentry-sdk==2.13.0
setproctitle==1.3.3
Shapely==1.8.1.post1
shortuuid==1.0.13
smmap==5.0.1
soupsieve==2.6
sparse==0.15.4
SQLAlchemy==2.0.22
stable_baselines3==2.3.2
stack-data==0.6.3
subprocess32==3.5.4
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
timm==1.0.15
tinycss2==1.4.0
toml==0.10.2
tomli==2.0.2
torch==2.4.1
torchaudio==0.10.2
tornado==6.4.1
tqdm==4.66.1
traitlets==5.11.2
triton==3.0.0
types-python-dateutil==2.9.0.20241003
tzdata==2023.3
uri-template==1.3.0
urllib3==2.2.2
wandb==0.12.7
wasabi==1.1.2
wcwidth==0.2.8
webcolors==24.8.0
Werkzeug==3.0.1
wrapt==1.15.0
xxhash==3.4.1
yarl==1.9.2
yaspin==2.5.0
zipp==3.17.0
zstandard==0.23.0
